Adjust rm statments.
svn path=/; revision=6423

+++ Systems/scripts/get_chill_image_cron.py
@@ -78,7 +78,8 @@ if len(regionlst) == 0:
     print "Must Exit!"
     # TODO need a nagios call here to alert operator 
     con.close()
-    os.remove(busy_file)
+    if os.path.isfile(busy_file):
+        os.remove(busy_file)
     sys.exit(1)
 region = (regionlst[0])[0]
 con.close()
